Remove panicking invocation of `transmute_copy` in `set` method
The unsafe use of `transmute_copy` has finally come home to roost, with
Rust 1.64.0 adding a check that the transmutation being performed is
legal, and panicking on failure.

While we're in here, let's clean up some brittle tests, document a lot
of items that are missing documentation, improve the homepage in the
cargo docs, and patch up some clippy lints.

Some of the changes necessary to fix the panic resulted in
backwards-incompatible changes, so I also bumped the version to 2.0.